Save and add OS functionality
Add OS functionality to the project by implementing thread creation and
starting. This change introduces two new threads, `blinky01` and
`blinky02`, which are started using the `OSThread_start()` function.
Additionally, the priorities of the PendSV and SysTick interrupts are
swapped to prevent preemption. The `BSP_SysInit()` function is also
called to initialize the system. The purpose of these changes is to
enable multi